% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjusted_hamming.R
\name{adjusted_hamming}
\alias{adjusted_hamming}
\title{Determine adjusted hamming}
\usage{
adjusted_hamming(x)
}
\arguments{
\item{x}{A binary matrix.}
}
\value{
A matrix of distances between each row of original matrix.
}
\description{
Fast calculation of Hamming distance when applied to a matrix.
}
\examples{
test_matrix <- matrix(c(0,1,1,
                      1,0,0,
                     0,1,1), ncol = 3, byrow = TRUE)
adjusted_hamming(test_matrix)
}
\references{
This fast Hamming distance version is based on: \url{https://johanndejong.wordpress.com/2015/09/23/fast-hamming-distance-in-r/}.
}
