% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply_weights.R
\name{apply_weights}
\alias{apply_weights}
\title{Apply weights}
\usage{
apply_weights(raw_data, weights)
}
\arguments{
\item{raw_data}{data.frame or matrix containing diagnostic data that is binary coded (1/0)}

\item{weights}{weights of the same length as the number of columns of combos}
}
\value{
A weighted matrix or data.frame with each position coded 1 changed to it's relevant weight.
}
\description{
Apply weights to a binary matrix or data.frame. Intended to be used after determining weights with \code{\link{get_prevalence_weights}}
}
\examples{

weights <- c(0.1, 0.2, 0.3, 0.4)
combos <- matrix(c(0,0,1,1,
                   1,0,1,1,
                   0,1,1,1), ncol=4, byrow=TRUE)

apply_weights(combos, weights)
}
