% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_prevalence_weights.R
\name{get_prevalence_weights}
\alias{get_prevalence_weights}
\title{Prevalence weighting of each variable}
\usage{
get_prevalence_weights(raw_data, cols)
}
\arguments{
\item{raw_data}{matrix or data.frame containing binary (0 or 1) data in tidy format.}

\item{cols}{Names or integer of column positions with relevant diagnosis}
}
\value{
list of different calculated prevalence measures TO BE UPDATED
}
\description{
Determine prevalence weights of each column of a data.frame or
}
\examples{
combos <- matrix(c(0,0,1,1,
                   1,0,1,1,
                   0,1,1,1), ncol=4, byrow=TRUE)

get_prevalence_weights(combos, 1:4)

}
