% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hamming.R
\name{hamming}
\alias{hamming}
\title{Determine Hamming distance}
\usage{
hamming(x)
}
\arguments{
\item{x}{A binary matrix.}
}
\value{
A matrix of distances between each row of original matrix.
}
\description{
Fast calculation of Hamming distance when applied to a matrix.
}
\examples{
test_matrix <- matrix(c(0,1,1,
                      1,0,0,
                     0,1,1), ncol = 3, byrow = TRUE)
hamming(test_matrix)
}
\references{
This fast Hamming distance version was taken from : https://johanndejong.wordpress.com/2015/09/23/fast-hamming-distance-in-r/
}
