% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hamming_prevalence.R
\name{hamming_prevalence}
\alias{hamming_prevalence}
\title{Prevalence weighted hamming}
\usage{
hamming_prevalence(raw_data, cols, weighting = NULL)
}
\arguments{
\item{raw_data}{matrix or data.frame containing binary coded data.}

\item{cols}{Names or integer of column positions; if not specified then all columns are used.}

\item{weighting}{Weighting measure to use passed as a string, options include \code{prevalence},
\code{inv_pr}, \code{inv_pr_wt}, \code{floored_pr}, or NULL TO BE UPDATED.}
}
\value{
A list with two items:
\describe{
  \item{adjusted_hamming}{A matrix of the hamming distance between each
  unique combinations in the provided \code{data}, weighted according to the specified \code{weighting}.
  This is a square matrix of dimension n*n}
  \item{key}{A vector of keys for each row of the returned \code{data},
  which is made up of the concatenated binary string across columns}.
  }
}
\description{
Calculate the
}
\examples{

dat <- matrix(c(0,0,1,1,
                   1,0,1,1,
                   0,1,1,1), ncol=4, byrow=TRUE)

hamming_prevalence(dat, 1:3)
}
